Make calls to `perf_event_open` read errors from errno
It looks like in the past this was not valid (see issue #4) but since
this was last changed things seem to have changed and now the sys
bindings for perf_event_open use libc::syscall which returns the error
in errno.

This commit replaces the check_raw_syscall calls around perf_event_open
with check_errno_syscall. This means that check_raw_syscall is no longer
used so it has been deleted. I have also includes a test case to verify
that the correct error code is returned in at least one case.

Fixes #18