Implement support for breakpoint events (#20)
* Combine `Event::r#type` and `Event::config` into `Event::update_attrs`

The previous split worked fine when the only fields that were worth
setting were `type` and `config`. However, once we get to other event
types like probes and breakpoints it will be necessary to set other
fields. Having a single method to do all of that works best.

* Add support for breakpoint events

* Add test cases that test both data and executable breakpoints

* Properly document `Breakpoint`

* Tweak doc wording

* Change event::Breakpoint to match review suggestions

* Remove file IO from test cases

* Formatting

* Make Breakpoint fn docs read better in search summaries.

All these descriptions began with "Create a breakpoint configuration
to...", but rustdoc searches show the beginning of each hit's first
doc paragraph, so it's nice to get the most distinctive information
about the function as early in the leading sentence as you can.

Co-authored-by: Jim Blandy <jimb@red-bean.com>