default FDs to CLOEXEC
Note: this is a behavior change.

Currently, this crate opens FDs without CLOEXEC, which means they
persist across exec. This is not the convention throughout the Rust
ecosystem, which is to do the opposite:

https://github.com/rust-lang/rust/issues/24237

This patch setting CLOEXEC :)

Note that doing that after-the-fact on the resulting FD isn't quite good
enough: if your process is forking a lot, that's racy.