Generate perf-event-open-sys bindings from kernel source
This commit contains two main changes:
- regenerate.sh has been rewritten to generate bindings from a local
  copy of the linux kernel source instead of using the headers installed
  on the current system.
- The bindings for x86 and arm64 have been regenerated from kernel
  version 6.2.10.

The benefit of this change is that it is easier to update the generated
bindings. It is not necessary to get a live install of whatever the
newest kernel version is to generate bindings for a new kernel version.
This is just barely doable with x86_64 and aarch64 but is rather
impractical for any one person to do once we think about potentially
bringing in other architectures. After this change, all that needs to be
done is update regenerate.sh with the new kernel verison and run it.