Remove Builder::group and replace it with Group::add
Builder::group means that the builder now has store a mutable reference
to a group. This affects the interface of a Builder (e.g. it can no
longer be cloned). A different approach, although not necessarily
better, is to add counters to a group. This change makes it so that
adding counters to a group now involves calling Group::add with the
Builder instance.