Allow user code to directly access Builder::attrs
At the moment this crate has decent coverage of the configuration
options within perf_event_attr. However, there will always be something
that we don't support, whether due to new kernel versions or for
options that we explicitly don't support. In these cases, providing
direct access to the underlying struct means that this crate continues
to be useful.

This commit adds attrs() and attrs_mut() methods to builder that
directly expose its internal perf_event_attr struct. The rest of the
changes are ensuring that user code mucking around with attrs doesn't
breake any of the assumptions of this crate.