Add a custom error type for when perf_event_open returns E2BIG
When one directly calls perf_event_open and it returns E2BIG it will
update the perf_event_attr::size field with the size it was expecting.
This allows for code interacting with it to do a fallback more easily.
While I don't think it's worth directly supporting that in this crate,
it is still useful to allow retrieving that value if the user wants to.

This commit makes this possible by replacing the E2BIG std::io::Error
with a wrapper around a new UnsupportedOptionsError type. Now, if they
so choose, users can get at the inner error and read the expected size.

I will admit, this is a fairly niche use-case but it is nice to support
it.