Move all `Group` functionality into `Counter` (#1)
Move all Group functionality into Counter

This commit takes all the functionality that was previously unique to
Group and adds it to Counter. It then takes Group and makes it a thin
wrapper around Counter that basically just forwards its relevant
methods. It ended up getting a bit out of hand but these are the two
core changes.

Motivation
==========
As things are, Group is pretty limited. Sure, you can add new counters
to it and it is useful to read them as a group. However, as soon as you
want to configure Group in any way, you're stuck.

At the kernel level, groups are just counters. So, what if we just
merged Groups and Counters altogether? Then, all the existing
configuration for Counters would also just work for groups because they
are just counters. Unfortunately, it turns out that having Group as a
counter you don't really care about is quite nice. So instead of
removing Group, I brought it back, made it a thin wrapper around a
Counter, kept all the new group-oriented methods on Counter, and got
this commit.

Detailed Changes
================
- You can now pass ReadFormat::GROUP to a Builder and a Counter built
  with it will work.
- Counters can now be group leaders in addition to Groups
- Counter now has group-wide versions of enable, disable, and reset.
  These work on any group member.
- The existing read, read_full, and read_count_and_time methods on
  Counter continue to work even when ReadFormat::GROUP is specified
  during setup. Internally, they read the values for the whole group and
  extract only the one for the current counter. The reverse happens when
  using read_group without ReadFormat::GROUP being specified.
- Builder can now be used to construct Groups. Group::new continues to
  work as it did before.
- Documentation for Counter, Group, and Builder has been revised in
  places to match up with the new semantics.
- Some other code has been reorganized as well.