Avoid generating bindings for non-perf types and constants
The current way of generating bindings via bindgen creates entries for
any type or constant that is reachable when including the headers listed
in wrapper.h. In addition to the perf-related structs we care about,
this includes:
- A variety of unrelated kernel API types
- A bunch of constants used as implementation details for making ioctls
- A constant for every single syscall number
- ... and more besides

There is no reason to include these in the generated bindings. Once
present, however, removing or changing them becomes a breaking change.
This is a one-shot change that changes the generation to only keep the
relevant structs.

This is a breaking change and would require a perf-event-open-sys2
v6.0.0.