Allow direct access to unnamed union fields on `perf_event_attr` (#48)
This means that future changes made to the C perf_event_attr struct
which replace existing fields with inline unions can continue to be
non-breaking changes to the crate.

Furthermore, it is generally much more convenient to directly access the
relevant fields instead of needing to go through the __bindgen_anon_x
field first.